+++ tests/communication/CMakeLists.txt
@@ -8,14 +8,14 @@ include_directories(
     ${CMAKE_CURRENT_SOURCE_DIR}/../testslave
 )
 
-add_executable(tst_modbuspoll ${CMAKE_CURRENT_SOURCE_DIR}/testmodbuspoll.cpp ${TEST_SRCS})
-add_executable(tst_communication ${CMAKE_CURRENT_SOURCE_DIR}/testcommunication.cpp ${TEST_SRCS})
-add_executable(tst_graphdatahandler ${CMAKE_CURRENT_SOURCE_DIR}/testgraphdatahandler.cpp ${TEST_SRCS})
-add_executable(tst_modbusconnection ${CMAKE_CURRENT_SOURCE_DIR}/testmodbusconnection.cpp ${TEST_SRCS})
-add_executable(tst_modbusmaster ${CMAKE_CURRENT_SOURCE_DIR}/testmodbusmaster.cpp ${TEST_SRCS})
-add_executable(tst_registervaluehandler ${CMAKE_CURRENT_SOURCE_DIR}/testregistervaluehandler.cpp ${TEST_SRCS})
-add_executable(tst_readregisters ${CMAKE_CURRENT_SOURCE_DIR}/testreadregisters.cpp ${TEST_SRCS})
-add_executable(tst_expressionparser ${CMAKE_CURRENT_SOURCE_DIR}/testexpressionparser.cpp ${TEST_SRCS})
+add_executable(tst_modbuspoll ${CMAKE_CURRENT_SOURCE_DIR}/tst_modbuspoll.cpp ${TEST_SRCS})
+add_executable(tst_communication ${CMAKE_CURRENT_SOURCE_DIR}/tst_communication.cpp ${TEST_SRCS})
+add_executable(tst_graphdatahandler ${CMAKE_CURRENT_SOURCE_DIR}/tst_graphdatahandler.cpp ${TEST_SRCS})
+add_executable(tst_modbusconnection ${CMAKE_CURRENT_SOURCE_DIR}/tst_modbusconnection.cpp ${TEST_SRCS})
+add_executable(tst_modbusmaster ${CMAKE_CURRENT_SOURCE_DIR}/tst_modbusmaster.cpp ${TEST_SRCS})
+add_executable(tst_registervaluehandler ${CMAKE_CURRENT_SOURCE_DIR}/tst_registervaluehandler.cpp ${TEST_SRCS})
+add_executable(tst_readregisters ${CMAKE_CURRENT_SOURCE_DIR}/tst_readregisters.cpp ${TEST_SRCS})
+add_executable(tst_expressionparser ${CMAKE_CURRENT_SOURCE_DIR}/tst_expressionparser.cpp ${TEST_SRCS})
 
 add_test(NAME tst_modbuspoll COMMAND tst_modbuspoll)
 add_test(NAME tst_communication COMMAND tst_communication)
