+++ tests/communication/tst_modbusmaster.cpp
@@ -6,7 +6,7 @@
 #include "testslavedata.h"
 #include "testslavemodbus.h"
 
-#include "testmodbusmaster.h"
+#include "tst_modbusmaster.h"
 
 #include <QMetaType>
