+++ tests/util/CMakeLists.txt
@@ -1,7 +1,7 @@
 
-add_executable(tst_updatenotify ${CMAKE_CURRENT_SOURCE_DIR}/testupdatenotify.cpp ${TEST_DATA_SRCS} ${GOOGLE_TEST_SOURCE})
-add_executable(tst_qmuparser ${CMAKE_CURRENT_SOURCE_DIR}/testqmuparser.cpp ${TEST_DATA_SRCS})
-add_executable(tst_formatrelativetime ${CMAKE_CURRENT_SOURCE_DIR}/testformatrelativetime.cpp ${TEST_DATA_SRCS})
+add_executable(tst_updatenotify ${CMAKE_CURRENT_SOURCE_DIR}/tst_updatenotify.cpp ${TEST_DATA_SRCS} ${GOOGLE_TEST_SOURCE})
+add_executable(tst_qmuparser ${CMAKE_CURRENT_SOURCE_DIR}/tst_qmuparser.cpp ${TEST_DATA_SRCS})
+add_executable(tst_formatrelativetime ${CMAKE_CURRENT_SOURCE_DIR}/tst_formatrelativetime.cpp ${TEST_DATA_SRCS})
 
 add_test(NAME tst_updatenotify COMMAND tst_updatenotify)
 add_test(NAME tst_qmuparser COMMAND tst_qmuparser)
